requirements.txt
asgi-redis==1.4.0
channels==1.1.3
Django==1.11.2
djangorestframework==3.6.3
django-allauth==0.32.0
django-debug-toolbar==1.6
django-environ==0.4.1
psycopg2==2.5.4